+++ Library_scheduler/include/CoCoPeLia.hpp
@@ -9,20 +9,20 @@
 #include "Autotuning_runtime.hpp"
 
 /// The CoCopeLia Dgemm implementation. A prediction model is used to select a tile from the micro-benchmarked tile candidates with CoCopeLia_optimize3.
-ATC_p CoCopeLiaDgemm(char TransA,  char TransB, size_t M, size_t N, size_t K,
-	double alpha, double* A, size_t ldA, double* B, size_t ldB, double beta, double* C, size_t ldC);
+ATC_p CoCopeLiaDgemm(char TransA,  char TransB, long int M, long int N, long int K,
+	double alpha, double* A, long int ldA, double* B, long int ldB, double beta, double* C, long int ldC);
 
 /// A modification of CoCopeLiaDgemm but with a given T (mainly for performance/debug purposes)
-ATC_p CoCopeLiaDgemmControled(char TransA,  char TransB, size_t M, size_t N, size_t K,
-	double alpha, double* A, size_t ldA, double* B, size_t ldB, double beta, double* C, size_t ldC, ATC_p predef_control_values);
+ATC_p CoCopeLiaDgemmControled(char TransA,  char TransB, long int M, long int N, long int K,
+	double alpha, double* A, long int ldA, double* B, long int ldB, double beta, double* C, long int ldC, ATC_p predef_control_values);
 
 /// The CoCopeLia Dgemm implementation. A prediction model is used to select a tile from the micro-benchmarked tile candidates with CoCopeLia_optimize3.
-ATC_p CoCopeLiaDaxpy(size_t N, VALUE_TYPE alpha,
-	VALUE_TYPE* x, size_t incx, VALUE_TYPE* y, size_t incy);
+ATC_p CoCopeLiaDaxpy(long int N, VALUE_TYPE alpha,
+	VALUE_TYPE* x, long int incx, VALUE_TYPE* y, long int incy);
 
 /// A modification of CoCopeLiaDgemm but with a given T (mainly for performance/debug purposes)
-ATC_p CoCopeLiaDaxpyControled(size_t N, VALUE_TYPE alpha,
-	VALUE_TYPE* x, size_t incx, VALUE_TYPE* y, size_t incy, ATC_p predef_control_values);
+ATC_p CoCopeLiaDaxpyControled(long int N, VALUE_TYPE alpha,
+	VALUE_TYPE* x, long int incx, VALUE_TYPE* y, long int incy, ATC_p predef_control_values);
 
 ///Deallocates the GPU-allocated cache buffer at target device
 void CoCopeLiaDevCacheFree(short dev_id);
